general reorganizing, the logging feature flag's not working lol
